Align benchmarks with refactor of thresholding to segmentation
