+++ benchmarks/benchmark_segmentation.py
@@ -150,3 +150,75 @@ class Watershed:
 
     def peakmem_watershed(self, seed_count, connectivity, compactness):
         watershed(self.image, seed_count, connectivity, compactness=compactness)
+
+
+class MultiOtsu:
+    """Benchmarks for MultiOtsu threshold."""
+
+    param_names = ['classes']
+    params = [3, 4, 5]
+
+    def setup(self, *args):
+        self.image = data.camera()
+
+    def time_threshold_multiotsu(self, classes):
+        segmentation.threshold_multiotsu(self.image, classes=classes)
+
+    def peakmem_reference(self, *args):
+        """Provide reference for memory measurement with empty benchmark.
+
+        Peakmem benchmarks measure the maximum amount of RAM used by a
+        function. However, this maximum also includes the memory used
+        during the setup routine (as of asv 0.2.1; see [1]_).
+        Measuring an empty peakmem function might allow us to disambiguate
+        between the memory used by setup and the memory used by target (see
+        other ``peakmem_`` functions below).
+
+        References
+        ----------
+        .. [1]: https://asv.readthedocs.io/en/stable/writing_benchmarks.html#peak-memory
+        """
+        pass
+
+    def peakmem_threshold_multiotsu(self, classes):
+        segmentation.threshold_multiotsu(self.image, classes=classes)
+
+
+class ThresholdLocalSauvola:
+    """Benchmark for transform routines in scikit-image."""
+
+    def setup(self):
+        self.image = np.zeros((2000, 2000), dtype=np.uint8)
+        self.image3D = np.zeros((30, 300, 300), dtype=np.uint8)
+
+        idx = np.arange(500, 700)
+        idx3D = np.arange(10, 200)
+
+        self.image[idx[::-1], idx] = 255
+        self.image[idx, idx] = 255
+
+        self.image3D[:, idx3D[::-1], idx3D] = 255
+        self.image3D[:, idx3D, idx3D] = 255
+
+    def time_sauvola(self):
+        segmentation.threshold_local_sauvola(self.image, window_size=51)
+
+    def time_sauvola_3d(self):
+        segmentation.threshold_local_sauvola(self.image3D, window_size=51)
+
+
+class ThresholdLi:
+    """Benchmark for threshold_li in scikit-image."""
+
+    def setup(self):
+        try:
+            self.image = data.eagle()
+        except ValueError:
+            raise NotImplementedError("eagle data unavailable")
+        self.image_float32 = self.image.astype(np.float32)
+
+    def time_integer_image(self):
+        segmentation.threshold_li(self.image)
+
+    def time_float32_image(self):
+        segmentation.threshold_li(self.image_float32)
